refactor: separate db working methods to service
